CSL: a checkin following a few days of debugging and checking on 32 bit platforms. As well aws some garbage collection error there I have improved time recording where previously if garbage collection happened during loading of a module the time for that GC was double accounted. I have enhanced my_assert to display clearer messages about failures when it needs to report them. And there is a new option "--kmax" which is in fact identical in behaviour to "--maxmem" but maybe fits alongside "-k" better. "--maxmem" is still available - and now might actually achieve what it is supposed to do.
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6702 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ csl/cslbase/syscsl.h
@@ -353,9 +353,8 @@ extern int terminal_eof_seen;
 // The next two call-backs tell the window manager how much time has
 // been spent and how many garbage collections have been done.
 //
-extern void report_time(int32_t t, int32_t gct);
-extern void report_space(uint64_t gccount, double percent,
-                         double mbytes);
+extern void report_time(uint64_t t, uint64_t gct);
+extern void report_space(uint64_t gccount, double percent, double mbytes);
 
 // These can be used (without 100% reliability!) to check if memory
 // addresses are proper. They are only for use in desparate debugging
